Extract existing sanitization code.
